Extract directions.
