+++ SnakeController/CMakeLists.txt
@@ -5,11 +5,13 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR})
 set(SNAKE_SOURCES
     SnakeController.cpp
 	Body.cpp
+	Directions.cpp
 )
 set(SNAKE_HEADERS
     SnakeController.hpp
     SnakeInterface.hpp
     Body.hpp
+	Directions.hpp
 )
 add_library(${TARGET_NAME} STATIC ${SNAKE_SOURCES} ${SNAKE_HEADERS})
 target_link_libraries(${TARGET_NAME} DynamicEvents)
