+++ SnakeController/Directions.hpp
@@ -1,0 +1,6 @@
+namespace Snake {
+bool isHorizontal(Direction direction);
+bool isVertical(Direction direction);
+bool isPositive(Direction direction);
+bool perpendicular(Direction dir1, Direction dir2);
+} // namespace
